requirements.txt
transformers
torch
requests
numpy
pandas
pyttsx3
pyaudio 
whisper 
elevenlabs
pyautogui
pytesseract
pillow
pygetwindow
pvporcupine
cryptography
pytest
pytest-cov
pytest-mock
pytest-asyncio
